Datasets/evalplus/Java/Code/HumanEval_27.java
package com.example;
public class HumanEval_27 {
    public static void main(String[] args) {
        System.out.println(flipCase("Hello"));
    }

    public static String flipCase(String str) {
        char[] charArray = str.toCharArray();
        for (int i = 0; i < charArray.length; i++) {
            if (Character.isUpperCase(charArray[i])) {
                charArray[i] = Character.toLowerCase(charArray[i]);
            } else if (Character.isLowerCase(charArray[i])) {
                charArray[i] = Character.toUpperCase(charArray[i]);
            }
        }
        return new String(charArray);
    }
}
// End of Code
