Datasets/evalplus/Java/src/test/java/com/example/HumanEval_27Test.java
package com.example;
import static org.junit.Assert.*;
import org.junit.Test;

public class HumanEval_27Test {
    @Test
    public void testFlipCase1() {
        assertEquals("", HumanEval_27.flipCase(""));
    }
    
    @Test
    public void testFlipCase2() {
        assertEquals("hELLO!", HumanEval_27.flipCase("Hello!"));
    }
    
    @Test
    public void testFlipCase3() {
        assertEquals("tHESE VIOLENT DELIGHTS HAVE VIOLENT ENDS", HumanEval_27.flipCase("These violent delights have violent ends"));
    }
    
    @Test
    public void testFlipCase4() {
        assertEquals("123ABC", HumanEval_27.flipCase("123abc"));
    }
    
    @Test
    public void testFlipCase5() {
        assertEquals("кАРЛ У кЛАРЫ УКРАЛ КОРАЛЛЫ, А к ЛАРА У кУАРЛА УКРАЛА КЛАРНЕТTtHE qUICK bROWN fox jumpΜΧΗΙs oVER THE LAZY DOGএএটি একটি ণউদাহরণক্T্ত কPমO CO.্ষেত্রhEsÍ,ЛАРА", HumanEval_27.flipCase("Карл у Клары украл кораллы, а К лара у Куарла украла кларнетtThe Quick Brown FOX JUMPμχῃS Over the lazy dogএএটি একটি ণউদাহরণক্t্ত কpমo co.্ষেত্রHeSí,лара"));
    }
    
    @Test
    public void testFlipCase6() {
        assertEquals("REкАРЛ У кЛАРЫ УКРАЛ КОРАЛЛЫ, А ", HumanEval_27.flipCase("reКарл у Клары украл кораллы, а "));
    }
}
// End of Test
